[ARM PATCH] 2188/1: Add missing MODULE_LICENSE declaration in PXA Lubbock PCMCIA driver
Patch from George G. Davis

The Lubbock PCMCIA socket driver, drivers/pcmcia/pxa2xx_lubbock.c,
lacks a MODULE_LICENSE declaration. This patch merely adds the missing
MODULE_LICENSE declaration to eliminate tainted kernel warnings when
this driver is built as a module. I have intentionally not used the
Dual MPL/GPL license in this case because 1) current linux-2.6.10-rc1
PXA PCMCIA drivers use only the GPL license, 2) the original source
files on which the PXA PCMCIA drivers are based, i.e. the SA11xx PCMCIA
drivers, state: "If you wish to allow the use of your version of this
file only under the terms of the GPL and not to allow others to use your
version of this file under the MPL, indicate your decision by deleting
the provisions above and replace them with the notice and other
provisions required by the GPL."


Signed-off-by:: George G. Davis