kbuild: Prefer Kbuild as name of the kbuild files
The kbuild syntax is unique and does only have very few things in common with
usual Makefile syntax. So to avoid confusion make the filename 'Kbuild' be
the preferred name as replacement for 'Makefile'.
No global renaming planned to take place for now, but new stuff expected to use
the new 'Kbuild' filename.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>