[ACPI] ACPICA 20041105 from Bob Moore
Implemented support for FADT revision 2.  This was an
interim table (between ACPI 1.0 and ACPI 2.0) that adds
support for the FADT reset register.

Implemented optional support to allow uninitialized LocalX
and ArgX variables in a control method.  The variables
are initialized to an Integer object with a value
of zero.  This support is enabled by setting the
acpi_gbl_enable_interpreter_slack flag to TRUE,
which is default unless booted with "acpi=strict".

Implemented support for Integer objects for the SizeOf
operator.  Either 4 or 8 is returned, depending on the
current integer size (32-bit or 64-bit, depending on the
parent table revision).

Fixed a problem in the implementation of the SizeOf and
ObjectType operators where the operand was resolved to
a value too early, causing incorrect return values for
some objects.

Fixed some possible memory leaks during exceptional conditions.