[PATCH] uml: re-add linux Makefile target - fixes to the old version.
We have re-added the "linux" target which just creates a hardlink to vmlinux. 
It was noted (sorry for forgetting who said this) that "make ARCH=um" will
leave an old "linux" file, since the target is not listed in .PHONY.

Using a symlink, instead of an hard link, was suggested because the hard link
does not get updated with make vmlinux.  But mv linux $destpath (the most
meaningful installation command) is broken, so I instead added linux to
.PHONY.  Obviously, linux will not be rebuilt on "make vmlinux ARCH=um", but
this makes sense (if people use the new way, don't use old tricks).  Anyway,
we remove "linux" in any case.

Signed-off-by: Paolo 'Blaisorblade' Giarrusso <blaisorblade_spam@yahoo.it>
Acked-by: Jeff Dike <jdike@addtoit.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>