[PATCH] md: "Faulty" personality
The 'faulty' personality provides a layer over any block device in which
errors may be synthesised.

A variety of errors are possible including transient and persistent read
and write errors, and read errors that persist until the next write.

There error mode can be changed on a live array.

Accessing this personality requires mdadm 2.8.0 or later.

Signed-off-by: Neil Brown <neilb@cse.unsw.edu.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>