[PATCH] cdev_init: zero out cdev before kobject_init()
Right now, cdev_init() works in a way that is not very intuitive.  If a
driver passes an uninitialized struct cdev to cdev_init(), then an
uninitialized struct kobject will be passed to kobject_init(), which does
kset_get() on kobj->kset, which probably points off into space and causes
an oops.  Drivers can work around this by zeroing out their struct cdev in
advance (and indeed most if not all of the things passed to cdev_init()
come from BSS) but I think it makes more sense for cdev_init() to live up
to its name and actually work on an uninitialized cdev.

Signed-off-by: Roland Dreier <roland@topspin.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>